+++ src/analyze/+mag/@IMAPTestingAnalysis/loadIALiRTData.m
@@ -35,8 +35,8 @@ function loadIALiRTData(this, primaryMetaData, secondaryMetaData)
         primary.compression = false(height(primary), 1);
         secondary.compression = false(height(secondary), 1);
 
-        primary.quality = true(height(primary), 1);
-        secondary.quality = true(height(secondary), 1);
+        primary.quality = repmat(mag.meta.Quality.Regular, height(primary), 1);
+        secondary.quality = repmat(mag.meta.Quality.Regular, height(secondary), 1);
 
         % Current file meta data.
         pmd = primaryMetaData.copy();