+++ resources/ReleaseNotes.md
@@ -6,6 +6,8 @@
 
 # Software
 
+- Add fine-grained definitions of quality flags (see `mag.meta.Quality`)
+- Update previous uses of quality flag
 - Allow cropping of `mag.Science` with negative duration (filters from the end)
 - Allow cropping of I-ALiRT with separate filter for primary and secondary
 - Add conversion method `eventtable` in `mag.event.Event` to simplify logic in science analysis